refactor: lotto가 보너스를 가졌는지 확인하는 메서드 추가, 테스트 추가
